Fix title screen, better settings, etc.
The title screen has better looking graphics for GUI focusing, and the anims now play properly.

Began implementing the settings API, for easier and universal usage of settings files in the codebase.

Also created a singletons folder, for singletons (duh), and moved SpaceGlobals into it. More will be moved into it in the future.