Reintroduced runtime sprite calculation
I've touched up the old sprite system to reuse it. Now, the palettes are stored using Godot's native Color variant. This means the magic 272727 number for transparency is no longer a thing (because it doesn't have to be)

I have also, with that, added in a new rendering function that can take any compressed sprite data, sprite dimensions, and palette, and return it as a Godot Image class.

The use of these sprites have been reimplemented for the ship and enemy, and the ship can load which sprite it's using based on the settings file attached to it, like planned.

I also took the chance to clean up a lot of the asset files, so there is now a slightly cleaner codebase :tada: